requirements.txt
graphql-core>=2<3
cached-property==1.4

# optional
django==2.0
psycopg2-binary>=2.7<3

# testing
pytest
pytest-cov
pytest-django

# QA
flake8
flake8-comprehensions
black

# documentation
sphinx
recommonmark  # markdown support
sphinx-rtd-theme
